idea for cyber securty?
I feel that bit coin is going to be regulated or killed by the US government sometime or later. So what would work to prevent cyber theft and or hacking from stealing bit coins? 

Could you apply regulations as those applied to credit cards to bit coin. 

Should you have a hot plug alert where people can mark wallets as robbed to track where their money went and to use local alerts to authorities or to sue potential cyber crimes by recipients of stolen coins? 

If a wallet is hacked should large transaction require a second verify by the owner of said wallet with a personal crypto id signature independent of the program and wallet profile?

Since criminals may pay in bit coins should markers send cyber alerts to Government or FBI that tells them bit coins were used to buy plutonium, nuclear materials, things for making bio weapons and weapons for mass destruction if that person is not a government or a legitimate organization? 

Can you mark transactions they would be related to weapons and stock piling for weapons for mass destruction in production or can you mark transactions that are used to buy drugs such as crystal meth and grey death and other narcotics? 

Marked transactions go to the proper government authorities. 

Then you have other transaction that may appear suspect, should markets for bit coins be regulated as such? And what would be the government protocols for prevention of cyber crime. Or would the US government even want an active participation in cyber policing transactions that are suspect?